requirements.txt
PySide6>=6.7
python-docx>=1.1.2
openpyxl>=3.1.5
pydantic>=2.7
pywin32>=306
pytest>=8.2
